project 2/2023010939/garbage_classification/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy
matplotlib
scikit-learn
tqdm
Pillow
opencv-python
pytorch-grad-cam